// prisma/generated/zod/inputTypeSchemas/AppInstallScopeCountOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const AppInstallScopeCountOrderByAggregateInputSchema: z.ZodType<Prisma.AppInstallScopeCountOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  appInstallId: z.lazy(() => SortOrderSchema).optional(),
  scopeId: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default AppInstallScopeCountOrderByAggregateInputSchema;